refactored note components into a subfolder of read
